game with broken lvls
many errors
