+++ Beeautiful/Beeautiful/Beeautiful/Beeautiful.csproj.Debug.cachefile
@@ -1,11 +1,12 @@
 Content\score.xnb
+Content\spritefonts\score.xnb
 Content\backgroundColor.xnb
 Content\backgroundClouds.xnb
 Content\speedLine.xnb
-Content\spritefonts\score.xnb
 Content\Sprites\JerryCenter\Jerry_Center.xnb
 Content\Sprites\JerryLeft\Jerry_left.xnb
 Content\Sprites\JerryRight\Jerry_right.xnb
+Content\background_level_1.xnb
 Content\Sprites\bg\backgroundClouds.xnb
 Content\Sprites\bg\backgroundColor.xnb
 Content\Sprites\bg\background_level_1.xnb
@@ -17,20 +18,19 @@ Content\Sprites\Jerry_Bee.xnb
 Content\Sprites\Jerry_bee_stinger.xnb
 Content\Sprites\Jerry_lives_death.xnb
 Content\Sprites\Jerry_lives_levelcomplete.xnb
-Content\Sprites\life.xnb
+Content\Sprites\stingGreenShot.xnb
+Content\Sprites\stingRedShot.xnb
+Content\Sprites\beatleBig.xnb
+Content\Sprites\beatleSmall.xnb
 Content\Sprites\shield.xnb
 Content\Sprites\smallcloud.xnb
 Content\Sprites\speedLine.xnb
 Content\Sprites\starBig.xnb
 Content\Sprites\starSmall.xnb
-Content\Sprites\stingGreenShot.xnb
-Content\Sprites\stingRedShot.xnb
-Content\Sprites\beatleBig.xnb
-Content\Sprites\beatleSmall.xnb
-Content\Audio\Map1.xnb
-Content\background_level_1.xnb
 Content\Jerry_lives_death.xnb
 Content\Sprites\lifeicon.xnb
 Content\Sprites\yellow_bee_enemy.xnb
 Content\Sprites\attacking_red_bee_enemy.xnb
+Content\Audio\Map1.xnb
+Content\Sprites\boss_placeholder.xnb
 Content\Audio\Map1.wma
